Added a settings manager
Added a settings manager since it shouldn't be inside the UI manager because it should only be for the Menu UI. The settings will be used in other places besides the Menu so it need its own script
